+++ src/lib/app-provider/tx/submitProposal.ts
@@ -2,7 +2,7 @@ import type { EncodeObject } from "@cosmjs/proto-signing";
 import type { StdFee } from "@cosmjs/stargate";
 import { useCallback } from "react";
 
-import { useCurrentChain } from "../hooks";
+import { useCurrentChain, useGetSigningClient } from "../hooks";
 import { trackTxSucceed } from "lib/amplitude";
 import {
   submitStoreCodeProposalTx,
@@ -20,7 +20,8 @@ export interface SubmitWhitelistProposalStreamParams {
 }
 
 export const useSubmitWhitelistProposalTx = () => {
-  const { address, getSigningCosmWasmClient } = useCurrentChain();
+  const { address } = useCurrentChain();
+  const getSigningClient = useGetSigningClient();
 
   return useCallback(
     async ({
@@ -31,7 +32,7 @@ export const useSubmitWhitelistProposalTx = () => {
       whitelistNumber,
       amountToVote,
     }: SubmitWhitelistProposalStreamParams) => {
-      const client = await getSigningCosmWasmClient();
+      const client = await getSigningClient();
       if (!address || !client)
         throw new Error("Please check your wallet connection.");
       if (!estimatedFee) return null;
@@ -49,7 +50,7 @@ export const useSubmitWhitelistProposalTx = () => {
         onTxFailed,
       });
     },
-    [address, getSigningCosmWasmClient]
+    [address, getSigningClient]
   );
 };